requirements.txt
Django==3.2.12
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
PyYAML==6.0
drf-yasg==1.20.0